i18n

+++ utils-support-common-starter/src/main/java/com/chua/common/support/utils/IdUtils.java
@@ -161,7 +161,7 @@ public class IdUtils {
      */
     public static String createUuid() {
         ThreadLocalRandom random = ThreadLocalRandom.current();
-        UUID uuid = new UUID(random.nextInt(), random.nextInt());
+        UUID uuid = new UUID(random.nextLong(), random.nextLong());
         return uuid.toString();
     }
